src/main/java/com/example/practise/interview/Interview.java
package com.example.practise.interview;

import java.util.HashMap;

public class Interview {

    public int sum(int a, int b) {
        return a + b;
    }

    public int[] twoSum(int[] nums, int target) {
        int[] result = new int[2];

        for (int i = 0; i < nums.length; i++) {
            int key = target - nums[i];

            for (int j = 1; j < nums.length; j++) {
                if (nums[i] + nums[j] == target) {
                    result[0] = i;
                    result[1] = j;
                    break;
                }
            }
        }
        return result;
    }


    public int[] twoSum1(int[] nums, int target) {
        int[] result = new int[2];
        HashMap<Integer, Integer> map = new HashMap();

        for (int i = 0; i < nums.length; i++) {
            int key = nums[i];
            map.put(key, i);
        }

        for (int i = 0; i < nums.length; i++) {
            Integer j = map.get(target - nums[i]);
            if (j != null) {
                result[0] = i;
                result[1] = j;
            }
        }
        return result;
    }


}
